ai_forecast/requirements.txt

psycopg2-binary==2.9.9
pandas==2.1.1
numpy==1.26.0
python-dotenv==1.0.1
statsmodels==0.14.0
